+++ packages/core/inline-edit/examples/00-basic-usage.js
@@ -1,7 +1,6 @@
 // @flow
 import React, { PureComponent } from 'react';
-import SingleLineTextInput from '@atlaskit/input';
-import InlineEdit from '../src';
+import InlineEdit, { SingleLineTextInput } from '../src';
 
 type State = {
   onEventResult: string,
@@ -35,7 +34,7 @@ export default class BasicExample extends PureComponent<void, State> {
 
   render() {
     return (
-      <div>
+      <div style={{ padding: '0 16px' }}>
         <InlineEdit
           label="Inline Edit Field"
           editView={