+++ packages/core/navigation-next/src/common/constants.js
@@ -4,5 +4,5 @@ export const transitionDuration = '0.22s';
 export const transitionDurationMs = 220;
 export const transitionTimingFunction = 'cubic-bezier(0.2, 0, 0, 1)';
 
-export const CONTENT_NAV_WIDTH = 270;
+export const CONTENT_NAV_WIDTH = 240;
 export const GLOBAL_NAV_WIDTH = 64;