+++ packages/core/navigation-next/src/components/GroupHeading/styles.js
@@ -1,8 +1,9 @@
 // @flow
 
-import { colors, gridSize } from '@atlaskit/theme';
+import { colors, gridSize as gridSizeFn } from '@atlaskit/theme';
 import type { ModeColors } from '../../theme/types';
 
+const gridSize = gridSizeFn();
 const fontSize = 11;
 
 const baseStyles = {
@@ -12,9 +13,9 @@ const baseStyles = {
     flexShrink: 0,
     fontSize: `${fontSize}px`,
     fontWeight: 600,
-    lineHeight: gridSize() * 2 / fontSize,
-    margin: `${gridSize() * 2.5}px 0 ${gridSize()}px`,
-    padding: `0 ${gridSize() / 2}px`,
+    lineHeight: gridSize * 2 / fontSize,
+    margin: `${gridSize * 4}px 0 ${gridSize}px`,
+    padding: `0 ${gridSize * 1.5}px`,
     textTransform: 'uppercase',
   },
   textWrapper: {
@@ -26,11 +27,16 @@ const baseStyles = {
   },
   afterWrapper: {
     lineHeight: 1,
-    marginLeft: `${gridSize() / 2}px`,
+    marginLeft: `${gridSize / 2}px`,
   },
 };
 
-export default ({ product }: ModeColors) => () => ({
+type GroupHeadingModeColors = {
+  product: any,
+  container: any,
+};
+
+export default ({ product }: ModeColors) => (): GroupHeadingModeColors => ({
   container: {
     ...baseStyles,
     headingBase: { ...baseStyles.headingBase, color: colors.N200 },