+++ packages/core/navigation-next/src/components/LayoutManager/types.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { ComponentType, Node } from 'react';
+import type { ComponentType, ElementRef, Node } from 'react';
 
 import UIController from '../../ui-controller/UIController';
 
@@ -21,11 +21,22 @@ export type CollapseListeners = {
   onCollapseEnd?: CollapseListener,
 };
 
+type NonStringRef<T> = {
+  current: ElementRef<T>,
+} | null;
+
+export type GetRefs = ({
+  expandCollapseAffordance: NonStringRef<'button'>,
+}) => void;
+
 export type ConnectedLayoutManagerProps = CollapseListeners & {
   /** Your page content. */
   children: Node,
   /** A component which will render the container navigation layer. */
   containerNavigation: ?ComponentType<{}>,
+  /** A function to access the refs of some elements within the LayoutManager
+   * component. */
+  getRefs?: GetRefs,
   /** A component which will render the global navigation bar. */
   globalNavigation: ComponentType<{}>,
   /** A component which will render the product navigation layer. */