+++ packages/editor/adf-utils/src/builders/nodes/doc.ts
@@ -1,25 +1,3 @@
-import {
-  ApplicationCardDefinition,
-  BlockQuoteDefinition,
-  BodiedExtensionDefinition,
-  BulletListDefinition,
-  CodeBlockDefinition,
-  DecisionItemDefinition,
-  DecisionListDefinition,
-  ExtensionDefinition,
-  HeadingDefinition,
-  MediaGroupDefinition,
-  MediaSingleDefinition,
-  OrderedListDefinition,
-  PanelDefinition,
-  ParagraphDefinition,
-  RuleDefinition,
-  TableDefinition,
-  TaskItemDefinition,
-  TaskListDefinition,
-  BlockCardDefinition,
-} from '@atlaskit/editor-common';
-
 import { DocNode, TopLevel } from '@atlaskit/editor-common';
 
 export const doc = (...content: TopLevel): DocNode => ({
@@ -27,25 +5,3 @@ export const doc = (...content: TopLevel): DocNode => ({
   version: 1,
   content,
 });
-
-export {
-  ApplicationCardDefinition,
-  BlockQuoteDefinition,
-  BodiedExtensionDefinition,
-  BulletListDefinition,
-  CodeBlockDefinition,
-  DecisionItemDefinition,
-  DecisionListDefinition,
-  ExtensionDefinition,
-  HeadingDefinition,
-  MediaGroupDefinition,
-  MediaSingleDefinition,
-  OrderedListDefinition,
-  PanelDefinition,
-  ParagraphDefinition,
-  RuleDefinition,
-  TableDefinition,
-  TaskItemDefinition,
-  TaskListDefinition,
-  BlockCardDefinition,
-};