+++ packages/editor/adf-utils/src/builders/nodes/panel.ts
@@ -1,11 +1,4 @@
-import {
-  BulletListDefinition,
-  HeadingDefinition,
-  OrderedListDefinition,
-  ParagraphDefinition,
-  PanelDefinition,
-  PanelAttributes,
-} from '@atlaskit/editor-common';
+import { PanelDefinition, PanelAttributes } from '@atlaskit/editor-common';
 
 export const panel = (attrs: PanelAttributes) => (
   ...content: PanelDefinition['content']
@@ -14,10 +7,3 @@ export const panel = (attrs: PanelAttributes) => (
   attrs,
   content,
 });
-
-export {
-  BulletListDefinition,
-  HeadingDefinition,
-  OrderedListDefinition,
-  ParagraphDefinition,
-};