+++ packages/editor/editor-core/example-helpers/mock-collab-provider.ts
@@ -155,7 +155,7 @@ export class MockCollabEditProvider implements CollabEditProvider {
   }
 }
 
-const getCollabEditProviderFor = <T>(participants: T) => (sid?: keyof T) =>
+const getCollabEditProviderFor = <T>(participants: T) => (sid?: string) =>
   Promise.resolve(new MockCollabEditProvider(mediator, sid));
 
 export const collabEditProvider = getCollabEditProviderFor(participants);