+++ packages/elements/analytics-listeners/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/analytics-listeners
 
+## 4.0.2
+- [patch] Update TS to 3.0 [f68d367](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f68d367)
+
 ## 4.0.1
 - [patch] fixed imports, docs and made GasPayload package attributes optional [6be5eed](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6be5eed)
 - [patch] use createAndFire function from analytics-next [095f356](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/095f356)